+++ CHANGELOG.md
@@ -32,6 +32,8 @@
     ([#235](https://github.com/microsoft/ApplicationInsights-Python/pull/235))
 - Fix export interval bug
     ([#237](https://github.com/microsoft/ApplicationInsights-Python/pull/237))
+- Defaulting logging env var for auto-instrumentation. Added logging samples.
+    ([#240](https://github.com/microsoft/ApplicationInsights-Python/pull/240))
 
 ## [1.0.0b8](https://github.com/microsoft/ApplicationInsights-Python/releases/tag/v1.0.0b8) - 2022-09-26